+++ CHANGELOG.md
@@ -41,6 +41,8 @@ Filter, sort and page parameters validation rules are excluded for resource requ
 parameters do not apply (create, read, update and delete).
 - [#92](https://github.com/cloudcreativity/laravel-json-api/issues/92)
 Last page link is now excluded if there are pages, rather than linking to page zero.
+- [#67](https://github.com/cloudcreativity/laravel-json-api/issues/67)
+Pagination meta will no longer leak into error response if error occurs when encoding data.
 
 ## [0.12.0] - 2018-02-08